+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 * After a savefile is renamed, flag the old savefile as removed
 * Support writes past the end of custom SaveFile streams
 * Only edit input editor analog values after double-click
+* Fix Vulkan bug where vkCreateInstance would not be hooked
 
 ## [1.4.6] - 2024-07-05
 ### Added